int to float data types

+++ Areas/SphereArea.py
@@ -5,5 +5,5 @@ def SphereArea(radius):
     return round(4 * pi * (pow(radius, 2)), 2)
 
 
-radius = int(input("Radius: "))
+radius = float(input("Radius: "))
 print(f"Area of the sphere's surface is {SphereArea(radius)}.")
